feat(heatmap): moved heatmap cesium sync to cesium
Moved implementation of HeatmapSynchronizer to Cesium as a simple workaround to handle the heatmap
canvas when swapping back and forth between renderer plugins

GEOVIZ-230
